feat: support Map for `CreateArray` literal
